+++ src/main/java/gift/controller/admin/AdminGlobalExceptionHandler.java
@@ -1,0 +1,26 @@
+package gift.controller.admin;
+
+import org.springframework.ui.Model;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import java.util.List;
+
+@ControllerAdvice
+public class AdminGlobalExceptionHandler {
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public String handleValidationExceptions(MethodArgumentNotValidException e, Model model) {
+        List<String> invalidReasonList = e.getBindingResult()
+                .getFieldErrors()
+                .stream()
+                .map(FieldError::getDefaultMessage)
+                .toList();
+        model.addAttribute("invalidReasons", invalidReasonList);
+
+        return "version-SSR/product-error";
+    }
+
+}
